+++ udata_piwik/client.py
@@ -11,6 +11,7 @@ from flask import current_app
 log = logging.getLogger(__name__)
 
 DEFAULT_TRACK_TIMEOUT = 60  # in seconds
+DEFAULT_ANALYZE_TIMEOUT = 60 * 5  # in seconds
 
 
 def analyze(method, **kwargs):
@@ -30,7 +31,9 @@ def analyze(method, **kwargs):
             dt = dt.isoformat()
         kwargs['date'] = dt
     data.update(kwargs)
-    r = requests.get(base_url, params=data)
+    timeout = current_app.config.get('PIWIK_ANALYZE_TIMEOUT',
+                                     DEFAULT_ANALYZE_TIMEOUT)
+    r = requests.get(base_url, params=data, timeout=timeout)
     return r.json()